+++ src/main/java/thunder/hack/modules/movement/Velocity.java
@@ -78,8 +78,8 @@ public class Velocity extends Module {
                         }
                     }
                     case Redirect -> {
-                        int vX = pac.getVelocityX();
-                        int vZ = pac.getVelocityZ();
+                        double vX = pac.getVelocityX();
+                        double vZ = pac.getVelocityZ();
                         if (vX < 0) vX *= -1;
                         if (vZ < 0) vZ *= -1;
                         double[] motion = MovementUtility.forward((vX + vZ));
